Multiple Mat / Obj sorting and match and folder creation
Added ability to add multiple materials to multiple objs and match them
up by name. Add to proper folder based on prefix or suffix.
Transform resetting, and adding materials to LODs,.

Other settings added for naming and such as well.

Useful for when you have a bunch of meshes imported that you have a lot
of swappable materials for, that you want to be able to apply all the
variations of the material to the objects based on names and create
prefabs for them in the appropriate folder.